[227-de-mo-generelle-nacharbeiten] add API to create tournament

+++ frontend/apps/desktop/src/app/pages/page-manage-team/team-bottom-bar/team-bottom-bar.component.ts
@@ -42,14 +42,10 @@ import { DialogModule } from 'primeng/dialog';
 })
 export class TeamBottomBarComponent {
   @Input() public form!: FormGroup<EditTeamForm>;
+  @Input() public teamId?: number;
   protected readonly ButtonColorEnum = ButtonColorEnum;
   protected readonly ButtonTypeEnum = ButtonTypeEnum;
   protected isDeleteDialogVisible = false;
-
-  @SingletonGetter()
-  public get team$(): Observable<TeamData | null> {
-    return this.store$.select(teamDetailsSelector);
-  }
 
   constructor(
     private readonly teamService: TeamService,
@@ -63,11 +59,13 @@ export class TeamBottomBarComponent {
   }
 
   public async onDeleteTeam() {
-    const teamId = (await firstValueFrom(this.team$))!.id;
+    if(!this.teamId) {
+      return;
+    }
 
     await firstValueFrom(
       this.teamService.delete({
-        teamId: teamId,
+        teamId: this.teamId,
       })
     );
 
@@ -76,28 +74,43 @@ export class TeamBottomBarComponent {
 
   public async onSaveTeam() {
     if (this.form.invalid) {
+      console.log(this.form);
       this.markAllFieldsAsTouched(this.form);
       return;
     }
 
-    const teamId = (await firstValueFrom(this.team$))!.id;
-
-    await firstValueFrom(
-      this.teamService.update({
-        teamId: teamId,
-        name: this.form.controls.name.value || undefined,
-        city: this.form.controls.city.value || undefined,
-        isMixTeam: this.form.controls.isMixTeam.value || undefined,
-        trainingTime: this.form.controls.trainingTime.value || undefined,
-        aboutUs: this.form.controls.aboutUs.value || undefined,
-        contacts:
-          this.form.controls.contacts.value.filter(
+    if(!this.teamId) {
+      console.log('here');
+      await firstValueFrom(
+        this.teamService.create({
+          name: this.form.controls.name.value,
+          city: this.form.controls.city.value || undefined,
+          isMixTeam: this.form.controls.isMixTeam.value,
+          trainingTime: this.form.controls.trainingTime.value || undefined,
+          aboutUs: this.form.controls.aboutUs.value || undefined,
+          contacts: this.form.controls.contacts.value.filter(
             (item): item is string => item !== null && item !== ''
           ) || [],
-      })
-    );
+        })
+      )
+    } else {
+      await firstValueFrom(
+        this.teamService.update({
+          teamId: this.teamId,
+          name: this.form.controls.name.value || undefined,
+          city: this.form.controls.city.value || undefined,
+          isMixTeam: this.form.controls.isMixTeam.value || undefined,
+          trainingTime: this.form.controls.trainingTime.value || undefined,
+          aboutUs: this.form.controls.aboutUs.value || undefined,
+          contacts:
+            this.form.controls.contacts.value.filter(
+              (item): item is string => item !== null && item !== ''
+            ) || [],
+        })
+      );
 
-    await this.teamDataService.reloadTeamDetails();
+      await this.teamDataService.reloadTeamDetails();
+    }
   }
 
   private markAllFieldsAsTouched(form: FormGroup): void {
